RMA-Demo/coordinator-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
pyjwt==2.8.0
